backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
celery
redis
authlib
python-jose[cryptography]
passlib[bcrypt]
alembic
httpx
python-multipart
itsdangerous
python-dotenv
